requirements.txt
jax
mujoco-mjx
wandb
einops
matplotlib
flax
moviepy
imageio